Creates separate PersonController and fills the file select field
