+++ src/main/resources/mainWindow.fxml
@@ -5,12 +5,12 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="240.0" prefWidth="420.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="UILoader">
+<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="240.0" prefWidth="420.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="PersonController">
    <center>
       <VBox alignment="CENTER" prefHeight="309.0" prefWidth="316.0" BorderPane.alignment="CENTER">
          <children>
             <Label text="File Location:" />
-            <TextField fx:id="fileText" />
+            <TextField fx:id="fileText" text="No Selection" />
             <Separator orientation="VERTICAL" prefHeight="0.0" prefWidth="39.0" visible="false" />
             <Button fx:id="browseBtn" mnemonicParsing="false" onAction="#runBrowseBtn" text="Browse" />
             <Separator orientation="VERTICAL" prefHeight="29.0" prefWidth="97.0" visible="false" />
